+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.6.0] - 2018-07-24
+### Removed
+- `facetsQuery` reference.
+
+### Added
+- multiselect for search filter.
+
 ## [0.5.2] - 2018-07-19
 ### Fixed
 - Remove code dependency with the resolved page path (e.g. `/eletronics/d`).